Fix: o nome dos diretorios foram modificados para melhor entendimento e conserto do diretorio gocosing para goconsig.
