+++ goconsig/app_core/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class BancosConfig(AppConfig):
+class CoreConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'bancos'
+    name = 'app_core'
