+++ goconsig/app_core/views.py
