+++ goconsig/app_parceiros/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class EmprestimosConfig(AppConfig):
+class ParceirosConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'emprestimos'
+    name = 'app_parceiros'
