+++ goconsig/app_usuarios/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ParceirosConfig(AppConfig):
+class UsuariosConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'parceiros'
+    name = 'app_usuarios'
