+++ goconsig/goconsig/asgi.py
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'gocosing.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'goconsig.settings')
 
 application = get_asgi_application()
