+++ goconsig/goconsig/urls.py
@@ -16,8 +16,8 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from core import views as core_views
-from usuarios import views as usuarios_views
+from app_core import views as core_views
+from app_usuarios import views as usuarios_views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
