+++ goconsig/templates/homes/home.html
@@ -3,14 +3,14 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>GoCosing</title>
+    <title>GoConsig</title>
     <link rel="icon" type="image/x-icon" href="">
     <link rel="stylesheet" href="">
 </head>
 <body>
     <header>
         <nav class="navbar">
-            <h1>Bem-vindo ao GoCosing</h1>
+            <h1>Bem-vindo ao GoConsig</h1>
             <div style="display: flex;" class="nav-links">
                 <a href="{% url 'login' %}">Login</a>
                 <p>ou</p>
@@ -23,7 +23,7 @@
         <p>Este é um exemplo de página inicial.</p>
     </main>
     <footer>
-        <p>&copy; 2025 GoCosing. Todos os direitos reservados.</p>
+        <p>&copy; 2025 GoConsig. Todos os direitos reservados.</p>
     </footer>
 </body>
 </html>
