+++ goconsig/templates/usuario/cadastro.html
@@ -3,13 +3,13 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>GoCosing</title>
+    <title>GoConsig</title>
     <link rel="icon" type="image/x-icon" href="">
     <link rel="stylesheet" href="">
 </head>
 <body>
     <header>
-        <h1>Bem-vindo ao GoCosing</h1>
+        <h1>Bem-vindo ao GoConsig</h1>
     </header>
     <main>
         <h2>Tela de Cadastro</h2>
@@ -18,7 +18,7 @@
         <a href="{% url 'login' %}">Login</a>
     </main>
     <footer>
-        <p>&copy; 2025 GoCosing. Todos os direitos reservados.</p>
+        <p>&copy; 2025 GoConsig. Todos os direitos reservados.</p>
     </footer>
 </body>
 </html>
